Remove dev/prod specific login.
This also removes the otherwise unused UnreachableCodeReachedError as
well as symbols or tests that reference dev login.
