+++ package.json
@@ -53,7 +53,6 @@
     "watch-css": "npm run build-css && node-sass-chokidar --include-path ./src src/ -o src/ --watch --recursive",
     "start-js": "react-scripts start",
     "start": "npm-run-all -p watch-css start-js",
-    "start-prod": "REACT_APP_LOGIN=prod npm start",
     "eslint": "eslint \"src/**/*.js\"",
     "lint": "prettier-eslint \"$PWD/src/**/*.js\"",
     "lint-fix": "prettier-eslint --write \"$PWD/src/**/*.js\"",
