+++ src/API/session.test.js
@@ -1,41 +1,6 @@
 import API from './index';
-import { mockHttpResponse, mockEnvironment } from '../TestHelper';
-import UnreachableCodeReachedError from '../Errors/UnreachableCodeReachedError';
+import { mockHttpResponse } from '../TestHelper';
 import ApiError from '../Errors/ApiError';
-
-describe('developmentLogin', () => {
-  const method = () => API.session.developmentLogin({ email: 'test' });
-
-  it('should throw Error in production', () => {
-    mockEnvironment.runInProduction(() => {
-      expect(method).toThrow(UnreachableCodeReachedError);
-    });
-  });
-
-  it('should return resolved promise on success', () => {
-    const response = { test: 'success' };
-    mockHttpResponse({
-      status: 200,
-      body: response
-    });
-
-    expect.assertions(1);
-
-    return method().then(res => expect(res).toEqual(response));
-  });
-
-  it('should return a rejected promise on failure', () => {
-    const response = { test: 'failure' };
-    mockHttpResponse({
-      status: 404,
-      body: response
-    });
-
-    expect.assertions(1);
-
-    return method().catch(err => expect(err).toBeInstanceOf(ApiError));
-  });
-});
 
 describe('login', () => {
   const method = () => API.session.login({ email: 'test', password: 'test' });
