--- src/Screens/Auth/Login/DevelopmentLogin/DevelopmentLogin.scss
@@ -1,5 +1,0 @@
-.development-login {
-  i {
-    display: block;
-  }
-}
